feat: Add CSV management for stations, tickets, and routes
- Implemented CSVManager class for handling CSV file operations.
- Added functionality to save and load stations and tickets from CSV files.
- Integrated CSV loading during system initialization to populate data.
- Updated analytics display to use color codes for better visibility.
- Enhanced ticketing process to persist ticket data to CSV.
- Added user authentication feature with a simple login mechanism.
- Refactored main menu and dashboard displays for improved user experience.
- Introduced new data files: routes.csv, tickets.csv, users.csv, and colors.h for ANSI color codes.
- Updated build scripts to compile new source files.

+++ include/scheduling.h
@@ -60,8 +60,8 @@ class MinHeap {
         int right = 2 * index + 2;
         int smallest = index;
 
-        if (left < heap.size() && heap[left] < heap[smallest]) smallest = left;
-        if (right < heap.size() && heap[right] < heap[smallest]) smallest = right;
+        if (left < (int)heap.size() && heap[left] < heap[smallest]) smallest = left;
+        if (right < (int)heap.size() && heap[right] < heap[smallest]) smallest = right;
 
         if (smallest != index) {
             std::swap(heap[index], heap[smallest]);
